Move notebook controls to drawer
