+++ src/components/Drawer/templates.jsx
@@ -1,5 +1,18 @@
-import React from 'react';
-import { Button, Switch, InputNumber, Select } from 'antd';
+/* eslint-disable react-refresh/only-export-components */
+import React, { useState, useEffect, useContext } from 'react';
+import {
+  Button,
+  Switch,
+  InputNumber,
+  Select,
+  Input,
+  Avatar,
+  Drawer as AntDrawer,
+} from 'antd';
+import { signOut } from 'next-auth/react';
+import { EditOutlined, UserOutlined } from '@ant-design/icons';
+import Link from 'next/link';
+import { ThemeContext } from '../ThemeProvider';
 
 /**
  * Template factory functions for Drawer.
@@ -96,7 +109,210 @@ export function editor({
   return { header, body };
 }
 
+function NotebookControllerContent({
+  onSelect,
+  showEdits,
+  onToggleEdits,
+  showArchived,
+  onToggleArchived,
+  onAddNotebookDrawerChange,
+}) {
+  const [notebooks, setNotebooks] = useState([]);
+  const [selected, setSelected] = useState('');
+  const [drawerOpen, setDrawerOpen] = useState(false);
+  const [newTitle, setNewTitle] = useState('');
+  const [newDescription, setNewDescription] = useState('');
+  const [newGroupAlias, setNewGroupAlias] = useState('');
+  const [newSubgroupAlias, setNewSubgroupAlias] = useState('');
+  const [newEntryAlias, setNewEntryAlias] = useState('');
+  const { darkMode, toggleTheme } = useContext(ThemeContext);
+
+  useEffect(() => {
+    async function fetchNotebooks() {
+      try {
+        const res = await fetch('/api/notebooks');
+        if (!res.ok) throw new Error('Failed to fetch notebooks');
+        const data = await res.json();
+        setNotebooks(data);
+        if (data.length) {
+          let initial = data[0].id;
+          if (typeof window !== 'undefined') {
+            const stored = localStorage.getItem('lastNotebookId');
+            if (stored && data.some((nb) => nb.id === stored)) {
+              initial = stored;
+            }
+            localStorage.setItem('lastNotebookId', initial);
+          }
+          setSelected(initial);
+          onSelect(initial);
+        }
+      } catch (err) {
+        console.error(err);
+      }
+    }
+    fetchNotebooks();
+  }, [onSelect]);
+
+  const handleChange = (e) => {
+    const id = e.target.value;
+    setSelected(id);
+    onSelect(id);
+    if (typeof window !== 'undefined') {
+      localStorage.setItem('lastNotebookId', id);
+    }
+  };
+
+  const handleCreate = async ({
+    name,
+    description,
+    user_notebook_tree,
+    precursorId,
+  }) => {
+    try {
+      const body = { title: name, description };
+      if (precursorId) {
+        body.precursorId = precursorId;
+      } else {
+        body.user_notebook_tree = user_notebook_tree;
+      }
+      const res = await fetch('/api/notebooks', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(body),
+      });
+      if (!res.ok) throw new Error('Failed to create notebook');
+      const newNb = await res.json();
+      setNotebooks((prev) => [...prev, newNb]);
+      setSelected(newNb.id);
+      setDrawerOpen(false);
+      onSelect(newNb.id);
+      if (typeof window !== 'undefined') {
+        localStorage.setItem('lastNotebookId', newNb.id);
+      }
+    } catch (err) {
+      console.error(err);
+    }
+  };
+
+  const handleCreateDrawer = async () => {
+    await handleCreate({
+      name: newTitle,
+      description: newDescription,
+      user_notebook_tree: [newGroupAlias, newSubgroupAlias, newEntryAlias],
+    });
+    setNewTitle('');
+    setNewDescription('');
+    setNewGroupAlias('');
+    setNewSubgroupAlias('');
+    setNewEntryAlias('');
+  };
+
+  useEffect(() => {
+    onAddNotebookDrawerChange?.(drawerOpen);
+  }, [drawerOpen, onAddNotebookDrawerChange]);
+
+  return (
+    <>
+      <div
+        style={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}
+      >
+        <select
+          value={selected}
+          onChange={handleChange}
+          style={{ width: '100%' }}
+        >
+          {notebooks.map((nb) => (
+            <option key={nb.id} value={nb.id}>
+              {nb.title}
+            </option>
+          ))}
+        </select>
+        <Button onClick={() => setDrawerOpen(true)} style={{ width: '100%' }}>
+          Add New
+        </Button>
+        <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
+          <Switch
+            checkedChildren={
+              <EditOutlined style={{ color: darkMode ? '#000' : '#fff' }} />
+            }
+            unCheckedChildren={
+              <EditOutlined style={{ color: darkMode ? '#fff' : '#000' }} />
+            }
+            checked={showEdits}
+            onChange={onToggleEdits}
+          />
+          <span>Manage</span>
+        </div>
+        <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
+          <Switch checked={showArchived} onChange={onToggleArchived} />
+          <span>Show Archived</span>
+        </div>
+        <Avatar
+          size="large"
+          icon={<UserOutlined />}
+          style={{ alignSelf: 'center' }}
+        />
+        <Link href="/account" style={{ marginBottom: '0.5rem' }}>
+          Account
+        </Link>
+        <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
+          <span>Dark Mode</span>
+          <Switch checked={darkMode} onChange={toggleTheme} />
+        </div>
+        <Button onClick={() => signOut({ redirect: false })}>Logout</Button>
+      </div>
+      <AntDrawer
+        title="New Notebook"
+        open={drawerOpen}
+        onClose={() => setDrawerOpen(false)}
+      >
+        <Input
+          placeholder="Name"
+          value={newTitle}
+          onChange={(e) => setNewTitle(e.target.value)}
+          style={{ marginBottom: '0.5rem' }}
+        />
+        <Input.TextArea
+          placeholder="Description (optional)"
+          value={newDescription}
+          onChange={(e) => setNewDescription(e.target.value)}
+          style={{ marginBottom: '0.5rem' }}
+        />
+        <Input
+          placeholder="Group Alias"
+          value={newGroupAlias}
+          onChange={(e) => setNewGroupAlias(e.target.value)}
+          style={{ marginBottom: '0.5rem' }}
+        />
+        <Input
+          placeholder="Subgroup Alias"
+          value={newSubgroupAlias}
+          onChange={(e) => setNewSubgroupAlias(e.target.value)}
+          style={{ marginBottom: '0.5rem' }}
+        />
+        <Input
+          placeholder="Entry Alias"
+          value={newEntryAlias}
+          onChange={(e) => setNewEntryAlias(e.target.value)}
+          style={{ marginBottom: '0.5rem' }}
+        />
+        <div style={{ display: 'flex', justifyContent: 'flex-end', gap: '0.5rem' }}>
+          <Button onClick={() => setDrawerOpen(false)}>Cancel</Button>
+          <Button type="primary" onClick={handleCreateDrawer}>
+            Create
+          </Button>
+        </div>
+      </AntDrawer>
+    </>
+  );
+}
+
+export function controller(props) {
+  return { body: <NotebookControllerContent {...props} /> };
+}
+
 export default {
   editor,
+  controller,
 };
