--- src/components/Menu/Menu.module.css
@@ -1,24 +1,0 @@
-.root {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  width: 100%;
-  height: 48px;
-}
-
-.left,
-.center,
-.right {
-  display: flex;
-  align-items: center;
-}
-
-.center {
-  flex: 1;
-  justify-content: center;
-}
-
-.right {
-  gap: 8px;
-}
